+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 2.81.0
 * Allow payee ID to be passed in options params for transaction create
 * Add `getMerchantId` to `ConnectedMerchantStatusTransitioned` and `ConnectedMerchantPayPalStatusChanged`, and `getOauthApplicationClientId` to OAuthAccessRevocation webhooks